Fix pass by value

+++ Exec/RegTests/PMF-SRK/prob.cpp
@@ -79,9 +79,11 @@ read_pmf(const std::string& myfile)
   PeleC::prob_parm_host->h_pmf_X.resize(PeleC::h_prob_parm_device->pmf_N);
   PeleC::prob_parm_host->pmf_X.resize(PeleC::h_prob_parm_device->pmf_N);
   PeleC::prob_parm_host->h_pmf_Y.resize(
-    PeleC::h_prob_parm_device->pmf_N * PeleC::h_prob_parm_device->pmf_M);
+    static_cast<long>(PeleC::h_prob_parm_device->pmf_N) *
+    PeleC::h_prob_parm_device->pmf_M);
   PeleC::prob_parm_host->pmf_Y.resize(
-    PeleC::h_prob_parm_device->pmf_N * PeleC::h_prob_parm_device->pmf_M);
+    static_cast<long>(PeleC::h_prob_parm_device->pmf_N) *
+    PeleC::h_prob_parm_device->pmf_M);
 
   iss.clear();
   iss.seekg(0, std::ios::beg);
